+++ package.json
@@ -39,6 +39,8 @@
     "webpack-dev-server": "^3.1.8"
   },
   "dependencies": {
-    "json-loader": "^0.5.7"
+    "history": "^4.7.2",
+    "json-loader": "^0.5.7",
+    "react-url-query": "^1.4.0"
   }
 }
